update build.sh and Readme.md

+++ build.sh
@@ -19,14 +19,14 @@ for file in ./docs/zh-CN/*.md; do
   echo '\n' >> "$ZH_CN_README";
 done;
 
-if which pretty_doc; then
-  pretty_doc -t parallel -o .
-else
-  echo 'You may want to install `pretty_doc` to generate static files:  `gem install pretty_doc`'
-fi
+# if which pretty_doc; then
+#   pretty_doc -t parallel -o .
+# else
+#   echo 'You may want to install `pretty_doc` to generate static files:  `gem install pretty_doc`'
+# fi
 
-mv README.html index.html
-mv README.en.html index.en.html
+if [ -f README.html ]; then mv README.html index.html; fi;
+if [ -f README.en.html ]; then mv README.en.html index.html; fi;
 
 sed -i "" "s/{:.no_toc}//g" README.md
 sed -i "" "s/+ toc-list/<!-- START doctoc -->/g" README.md
@@ -35,5 +35,5 @@ sed -i "" "s/{:toc}/<!-- END doctoc -->/g" README.md
 if which doctoc; then
   doctoc README.md
 else
-  echo 'You may want to install `doctoc` to generate TOC:  `npm install doctoc`'
+  echo 'You may want to install `doctoc` to generate TOC:  run `npm install -g doctoc`'
 fi
